Prevent migration BigbluebuttonRailsTo200E from breaking
Would break when migrating from 0.8.1.

+++ db/migrate/20141210192333_bigbluebutton_rails_to200_e.rb
@@ -1,9 +1,13 @@
 class BigbluebuttonRailsTo200E < ActiveRecord::Migration
   def up
-    remove_index :bigbluebutton_rooms, :voice_bridge
+    if index_exists?(:bigbluebutton_rooms, :voice_bridge)
+      remove_index :bigbluebutton_rooms, :voice_bridge
+    end
   end
 
   def down
-    add_index :bigbluebutton_rooms, :voice_bridge, :unique => true
+    unless index_exists?(:bigbluebutton_rooms, :voice_bridge)
+      add_index :bigbluebutton_rooms, :voice_bridge, unique: true
+    end
   end
 end
